requirements.txt
fastapi
uvicorn
python-multipart
torch
transformers
openai-whisper
nltk
pydantic
ffmpeg-python
